deep-mind: put our trace BEFORE the actual removal of the object, hoping that the memory isn't free'd here and we won't get corrupt data.
There are good chances that upon `DB_OPERATION REM`, the contents of
`obj` were already freed and potentially reallocated to some other schtufs.
If we trace it before the deletion, we should have integrity on the data.

More tests need to happen because we don't have traces of this happening
in our instrumentation.