First draft of a slimmer Docker CDT image
This is probably the last commit related to CDT in this package. Initial
support for Dockefile has landed, it's a question of time before this
package get hosted on Docker Hub. However, current proposed Dockerfile in
eosio.cdt is not on par with this one, see [PR #227](https://github.com/EOSIO/eosio.cdt/pull/227)
for further details.

The reason of adding g++ is also noted in the PR comments (quickly, it's
because a valid normal non cross-compiling C++ is required due to CMake).